🚚 채팅 컴포넌트별 분리로 인한 파일 생성 및 삭제
