+++ src/components/main/chat/chatContents/index.style.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const Contents = styled.div`
+    font-style: normal;
+    font-weight: 400;
+    font-size: 14px;
+    line-height: 18px;
+    padding-top: 100px;
+`;
