+++ src/components/main/chat/chatList/index.jsx
@@ -1,20 +1,10 @@
-import ChatListItem from './ChatListItem';
-import styled from 'styled-components';
-
-const Wrapper = styled.div`
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    justify-content: center;
-
-    & > * {
-        :not(:last-child) {
-        margin-bottom: 20px;
-        }
-    }
-`;
+import ChatListItem from '../chatListItem/index.jsx';
+import { Wrapper } from './index.style';
 
 function ChatList(props) {
+    // Chatting에서 props 받아옴
+    // ChatList는 각방(item)에 key, room을 props로 전달
+    // 채팅 각방(item)을 클릭하면 room이 나옴
     const { rooms, onClickItem } = props;
 
     return (
