+++ src/components/main/chat/chatListItem/index.style.js
@@ -1,24 +1,26 @@
 import styled from 'styled-components';
-import partnerProfileImage from '../../../assets/basic-profile-img.png';
 
-const Wrapper = styled.div`
+export const Wrapper = styled.div`
     width: 90vw;
     display: flex;
     flex-direction: row;
     margin: 0 auto;
     justify-content: center;
     padding: 5px;
-    border-radius: 5px;
+    border-radius: 10px;
     cursor: pointer;
+    &:hover {
+        background-color: #B2EBF2;
+    }
 `;
 
-const ChatRoomImg = styled.img`
+export const ChatRoomImg = styled.img`
     width: 42px;
     height: 42px;
     margin-right: 12px;
 `;
 
-const ChatPartnerName = styled.p`
+export const ChatPartnerName = styled.p`
     font-style: normal;
     font-weight: 800;
     font-size: 16px;
@@ -30,7 +32,7 @@ const ChatPartnerName = styled.p`
     text-overflow: ellipsis;
 `;
 
-const ChatPreview = styled.p`
+export const ChatPreview = styled.p`
     font-style: normal;
     font-weight: 400;
     font-size: 14px;
@@ -42,7 +44,7 @@ const ChatPreview = styled.p`
     text-overflow: ellipsis;
 `;
 
-const ChatDate = styled.p`
+export const ChatDate = styled.p`
     font-style: normal;
     font-weight: 400;
     font-size: 12px;
@@ -50,24 +52,3 @@ const ChatDate = styled.p`
     color: #DBDBDB;
     margin-top: 24px;
 `;
-
-function ChatListItem(props) {
-    const { room, onClick } = props;
-    
-    return (
-        <Wrapper onClick={onClick}>
-                <div>
-                    <ChatRoomImg src={partnerProfileImage} alt='' />
-                </div>
-
-                <div>
-                    <ChatPartnerName>{room.partner}</ChatPartnerName>
-                    <ChatPreview>{room.content}</ChatPreview>
-                </div>
-
-                <ChatDate>{room.date}</ChatDate>
-        </Wrapper>
-    );
-}
-
-export default ChatListItem;
